:art: 공용 api fetcher생성 및 api url 한변수에서 관리
