--- apps/web/src/apis/utils/onRequest.ts
@@ -1,10 +1,0 @@
-import { InternalAxiosRequestConfig } from "axios";
-
-export const onRequest = (
-  config: InternalAxiosRequestConfig,
-): InternalAxiosRequestConfig => {
-  const { method, url } = config;
-  console.log(`📣 [API - REQUEST] ${method?.toUpperCase()} ${url}`);
-
-  return config;
-};
