+++ apps/web/src/assets/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './fetcher';
